Increase control signal accuracy
Since the signal duration is measured in the main loop, it is affected by the loop run duration. This can vary significantly if the program switches LED program frames. But this happens only once and then. So the strategy is to discard control signals if they were measured during a heavy main loop iteration.